Bug: 476354031
Test: CI

+++ src/amd/vulkan/radv_image_view.c
@@ -19,8 +19,6 @@
 #include "radv_image_view.h"
 
 #include "ac_descriptors.h"
-#include "ac_formats.h"
-#include "gfx10_format_table.h"
 
 static unsigned
 radv_tex_dim(VkImageType image_type, VkImageViewType view_type, unsigned nr_layers, unsigned nr_samples,
@@ -428,7 +426,7 @@ radv_image_view_make_descriptor(struct radv_image_view *iview, struct radv_devic
 
       /* Video decode target uses custom height alignment. */
       if (image->vk.usage & VK_IMAGE_USAGE_VIDEO_DECODE_DST_BIT_KHR) {
-         offset += first_layer * image->planes[plane_id].surface.u.legacy.level[0].slice_size_dw * 4;
+         offset += first_layer * (uint64_t)image->planes[plane_id].surface.u.legacy.level[0].slice_size_dw * 4;
          first_layer = 0;
       }
    }
